requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
python-multipart>=0.0.9
python-dotenv>=1.0.1
requests>=2.32.3
boto3>=1.34.135
pydantic>=2.7.4
pillow>=10.3.0
instructor>=1.3.3
